Preload Scene & file data progress
Made preload scene where we will create the non-destroy-able GameObject that will host static variables for savedata and house preloaded audio files.
Also made progress on save data files on game start-up and game load. Also got the ability to check if the file has been tampered with, which will be fun to punish cheaters with.