Interactable objects started
Added to the LookAtDoor script where i could get interactable objects functionality a home here.
Rn have text showing up when approaching the door on the third startup attempt.
Would like to get it when you click e on the door the passcode UI appears and the player and camera is locked in place.