Loading into level 2 successful
Loading into level 2 successful. For puzzle 2 because it will be an in-depth puzzle and very interactive i will probs have another local controller in that scene itsel or add it to the gui prefab..
Also made it so the player is rotated correctly depending on the scene name.
Preload is going to be necessary for this game to keep from duplicate DoNotDestroy gameobjects from spawing.