Q&A startup
Have it so, like the door lock code mechanic the same happens for the mirror, but the mirror will pop-up with the Q&A questions instead.
TODO:
Make it so the player transitions to a center position in front of the mirror and rotates directly at it so the view and distance is always the same.
Right now you can click e to exit Q&A, i don't want the player to be able to do that.
Make it so you can actually answer the questions!!!