Answering questions works flawlessly
Able to answer questions and have in-game GUI react correctly and smoothly due to results.
Altered the experience with the formbars to make it more user-friendly, allowing user to click 'enter' to enter and having formbar automaticlly selected on unhide.
Altered the wording on some of the Questions in Database.txt.