Random Relic Spawn Locations made
Added a bunch of relic spawn GOs and made a script hide all of them and then unhide 15 random ones. This will be used to spawn relic prefabs on when the hunt begins and it will be random each time the player plays at this level.
Also added a slot in DoorOpen to trigger the relic hunt to begin, but did not implement it yet.
Also changed some folder names in the Scripts folder to make more sense.
ALSO: forgot to mention that I closed off certain areas of player travel so that player cannot hide from where the AI's Navmesh does not travel. This was not tested thoroughly.