Relic spawn progress
Made it so that unused RelicSpawnLocations Despawn and Relics spawn on the 15 chosen points.
Also changed DestroySP.cs to DestroyOBJ.cs because I will be using it for any GOs I want to destroy.
Will need to make it so the relics do not spawn until you attempt to open the basement door and then make AI unhidden or make him spawn instead since he is a prefab.