Lil bit of work on QandA
Got a hint light to trigger and the door to the basement to unlock after the completion of the QandA game.
Also fixed the glitch on the door passcode gui where clicking e would exit you out of it, it is now the esc key instead, so players can type e into the form.

Next I want to make the final scene of walking down the stairs, i will need an endless staircase so i will create a prefab for it and duplicate it c: