Critique Changes
Change chiller font on only a few UI elements to showcase what my backup font looks like.
Also altered the camera rotation values so it is not as exaggerated, still will need to change in the future.
And I got rid of a lot of the pop-culture questions, sad day...