Audio Progress
Have the AudioController script setup so that it will play the flash back hints randomly and not repeat them until you go through the entire list!

Stopped this audio on starting game buttons... hope i dont have to do this too much, I think it will be fine in-game on a timer.