Audio Progress
Have it so the audio now adjusts correctly with option sliders on both MM and in-game.
Added a bunch of safety nets for if you exit the game or pause, the static hint sounds will not play audio.
Also got it so if you open or close a door and pause it the audio will freezer then resume when unpaused.

This sucked more than I thought it would.