JumpScareTrigger works
JumpScareTrigger script now works better, actually functions how I wanted it too.
On trigger enter still triggers twice, but not too bad of results.
Also, I removed the "Player" tag from HidingCheck, which is a child gameobject of FPSController Prefab. As far as I know this was not needed as a tag and does not cause any known glitches. I found that removing the Player tag caused less extra calls for the jumpscare script. Hopefully this doesn't mess with the hiding in a wardrobe mechanics, I did test slightly, but we shall see.