Glitches and test build ready
Discovered 2 glitches and fixed them.
On game end or on exit to menu in relic hunt scene if you left before the hunt starts then an error occurs OnDisable in JumpscareHides.cs, I believe the glitch was solved by checking if the relic hunt script is active and enabled instead of running the script which tries to hide a list of Gameobjects that are unactive.
Also found the cause of the rare glitch that makes AI ignore you and does not allow you to interact with anything. It was the hiding in a waredrobe and the destruction of that waredrobe. Ishiding would be stuck to true which means you could not be seen by AI. Also since the LookAt script ties the current obj to the waredrobe you were hiding in, it would not allow the raycast to change the LookAt variable.
Believe it was fixed in AIMain.cs.