A lot: sprinting mechanic, exhaustion mechanic, some audio
merged and implemented clicking button noise for all UI except Q&A scene and scene story UI. Also as a note for typing in pwd I want either a jump-scare or a buzzing sound.
I also added the feature of the file slots to include current level of game and last file write date on the slot button text, I think this is a very nice feature!

Also added a sprinting feature for the player and added an exhaustion effect if stamina runs out. This can be viewed by the new added Slider to help the player keep track of their stamina, so they don't get exhausted. 
The values of how long you can sprint, max sprint time, speed, walk speed, and exhausted delayed timer may have to be adjusted in the future but they work and the player's speed and camera movement are adjusted to look and feel like you are running.