Q&A chase sequence
I got the post-processing and ambient audio portion of the chase sequence finished, just have to get the mini AI logic made and then make him spawn and de-spawn properly, but it shouldn't be hard because everything is all setup. Just have to get the AI made and jump-scare camera switch to work for this scenario too, but make game over happen if he gets you.
This will add a final intense chase scene to the game for a final scare.