Got final intense AI Chase scene
Very spooky and a good rush to the basement door. Good last intense finale, and it is working and fixed bugs with it.
Found one bug when the ambient audio wasnt changing after answering final question in QandA, it seemed to work well after i tested it, but i don't know if it will repeat or not.