Repaired NavMesh Glitches
Finished the Audio Story Scene, it is now finished with functional and informative UI and Gifs to help aid in the audio story.
Also I messed around with the Nav Mesh settings and was able to lower the voxels settings in the nav mesh bake, which made it so AI can now get closer to objects. This completely fixed the glitch of being able to stump the AI from getting to you if you hug certain objects.