Before Dinner
Night before class got:
Ambient sound change glitch fixed, when it was not changing properly I found my code that was ruining it.
Made the "Between" scene that plays the car accident audio before the start of the game, which is the protagonist reliving his traumatic event that he is guilty over before the simulation begins.
Got the HUDs to disappear while game is paused and reappear on Unpause.
Got the glitches fixed with the stamina bar HUD element aka: disappear when paused, not go down while not moving, and disappear while in Q&A.

Also for the Scene Audio Player I made it so the play button only appears at appropriate times.