After Dinner
Got the story clips in the beginning of the game and at the end of the game implemented, it was worse than I though...
Also shortened the amount of stairs you have to go down at the end to 5 spawns instead of 15. its more bearable.
Set up Slot1 as a test slot, you can manipulate the txt save data file to play whatever level you want, this won't trigger the anti cheater scene.