Bunch of Audio Added
Added a bunch of audio for the AI's voice, might add more later.
Added Breathing Audio and implemented it in-game.
The breathing audio only triggers if you are exhausted and the protagonist or flash back audio is not already playing. both of these will interrupt the breathing audio too for story purposes.
Also getting caught by the AI will stop the sound too.