Fixed Discovered Bugs
Got the raycast visual to not glitch when going into a wardrobe, it disappears.
Also got it to disappear at appropriate times like getting jump-scared by AI and at game over and game complete scenes.

Also made it so the AI Voices and footsteps are not played anymore on Game Over.