Lots of good stuff
Got the locked door sound effect in

Got the logic for story audio hints in, but now just gotta put it in certain places with a certain %'s and test it!

Fixed the ambient audio in the beginning of the game

Added a confirmation window to exiting game and exiting application.