Playtest Changes with HD
Got rid of some riddles and altered the answers for others, that HD and I thought would work as answers too. 

Discovered and fixed glitch where if you are moving and shut yourself in a wardrobe, the player is still bobbing up and down and making footstep noises.

altered the hint texts in the Intro level slightly to make ctrl v concept more clear.