Maybe Final Push
Not a huge deal, just change the nav mesh agent stopping distance and value in script for the AI animation triggers to make it a better chance of the glitch where the AI animation doesnt go to idle when searching and not moving. Will be implemented on build but not working files, which is fine it's not a huge bug.